fix: rerender optimization, camera bounds
